before deleting, check if the category or user has posts

+++ src/Controller/Admin/Blog/CategoryController.php
@@ -75,6 +75,12 @@ class CategoryController extends AbstractController
     #[Route('/{id}', name: 'admin_blog_category_delete', methods: ['POST'])]
     public function delete(Request $request, Category $category): Response
     {
+        if ($category->getPosts()->count()) {
+            $this->addFlash('warning', 'Category is not empty!');
+
+            return $this->redirectToRoute('admin_blog_category_index', [], Response::HTTP_SEE_OTHER);
+        }
+
         if ($this->isCsrfTokenValid('delete'.$category->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($category);
